+++ src/azure-cli/azure/cli/command_modules/role/tests/latest/test_role.py
@@ -305,8 +305,18 @@ class RoleAssignmentScenarioTest(RoleScenarioTest):
             self.cmd('ad user create --display-name tester123 --password Test123456789 --user-principal-name {upn}')
             time.sleep(15)  # By-design, it takes some time for RBAC system propagated with graph object change
 
+            base_dir = os.curdir
             try:
-                self.cmd('configure --default group=' + resource_group)
+                temp_dir = self.create_temp_dir()
+                os.chdir(temp_dir)
+                self.cmd('configure --default group={rg} --scope local')
+                local_defaults_config = self.cmd('configure --list-defaults --scope local', checks=[
+                    self.check('length([])', 1),
+                    self.check('[0].name', 'group'),
+                    self.check('[0].value', '{rg}')
+                ]).get_output_in_json()
+                self.assertTrue(temp_dir.lower() in local_defaults_config[0]['source'].lower())
+
                 # test role assignments on a resource group
                 rg_id = self.cmd('group show -n {rg}').get_output_in_json()['id']
                 self.cmd('role assignment create --assignee {upn} --role reader --scope ' + rg_id)
@@ -314,7 +324,8 @@ class RoleAssignmentScenarioTest(RoleScenarioTest):
                 self.cmd('role assignment delete --assignee {upn} --role reader --scope ' + rg_id)
                 self.cmd('role assignment list --assignee {upn} --role reader --scope ' + rg_id, checks=self.check('length([])', 0))
             finally:
-                self.cmd('configure --default group=""')
+                self.cmd('configure --default group="" --scope local')
+                os.chdir(os.path.basename(base_dir))
                 self.cmd('ad user delete --upn-or-object-id {upn}')
 
     @ResourceGroupPreparer(name_prefix='cli_role_assign')