+++ src/command_modules/azure-cli-eventgrid/azure/cli/command_modules/eventgrid/_client_factory.py
@@ -14,6 +14,14 @@ def topics_factory(cli_ctx, _):
     return cf_eventgrid(cli_ctx).topics
 
 
+def domains_factory(cli_ctx, _):
+    return cf_eventgrid(cli_ctx).domains
+
+
+def domain_topics_factory(cli_ctx, _):
+    return cf_eventgrid(cli_ctx).domain_topics
+
+
 def event_subscriptions_factory(cli_ctx, _):
     return cf_eventgrid(cli_ctx).event_subscriptions