+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -3,6 +3,8 @@
 Release History
 ===============
 
+* `policy assignment list`: Fix error when using a resource group or subscription level `--scope`.
+
 2.1.15
 ++++++
 * `deployment create`: Improve error message when there is no TTY available